+++ tests/test_reply.py
@@ -2,10 +2,10 @@
 
 import time
 from werobot.parser import parse_user_msg
-from werobot.replies import TextReply, SuccessReply
+from werobot.replies import TextReply, ImageReply, SuccessReply
 
 
-def test_render_text_reply():
+def test_text_reply():
     message = parse_user_msg("""
         <xml>
         <ToUserName><![CDATA[toUser]]></ToUserName>
@@ -26,7 +26,32 @@ def test_render_text_reply():
     <CreateTime>{time}</CreateTime>
     <MsgType><![CDATA[text]]></MsgType>
     <Content><![CDATA[aa]]></Content>
-    <FuncFlag>0</FuncFlag>
+    </xml>""".format(time=t).strip()
+
+
+def test_image_reply():
+    message = parse_user_msg("""
+        <xml>
+        <ToUserName><![CDATA[toUser]]></ToUserName>
+        <FromUserName><![CDATA[fromUser]]></FromUserName>
+        <CreateTime>1348831860</CreateTime>
+        <MsgType><![CDATA[image]]></MsgType>
+        <PicUrl><![CDATA[this is a url]]></PicUrl>
+        <MediaId><![CDATA[media_id]]></MediaId>
+        <MsgId>1234567890123456</MsgId>
+        </xml>
+    """)
+    t = int(time.time())
+    reply = ImageReply(message=message, media_id="fdasfdasfasd", time=t)
+    reply.render().strip() == """
+    <xml>
+    <ToUserName><![CDATA[fromUser]]></ToUserName>
+    <FromUserName><![CDATA[toUser]]></FromUserName>
+    <CreateTime>{time}</CreateTime>
+    <MsgType><![CDATA[image]]></MsgType>
+    <Image>
+    <MediaId><![CDATA[fdasfdasfasd]]></MediaId>
+    </Image>
     </xml>""".format(time=t).strip()